Fixed CA errors by moving classes to more appropriate CA layers and adding a new boundary class
